refactor: Typography 컨벤션에 맞게 수정 및 px에서 rem으로 수정
