+++ src/lib/foundation/Typography/styles.ts
@@ -2,74 +2,74 @@ import { styled } from 'styled-components';
 
 export const Headline = styled.h1`
   font-weight: 700;
-  font-size: 40px;
+  font-size: 4rem;
   line-height: 1.1;
 `;
 
 export const Title1 = styled.h2`
   font-weight: 700;
-  font-size: 38px;
+  font-size: 3.8rem;
   line-height: 1.1;
 `;
 
 export const Title2 = styled.h3`
   font-weight: 700;
-  font-size: 36px;
+  font-size: 3.6rem;
   line-height: 1.1;
 `;
 
 export const Title3 = styled.h4`
   font-weight: 700;
-  font-size: 32px;
+  font-size: 3.2rem;
   line-height: 1.28;
 `;
 export const Title4 = styled.h5`
   font-weight: 700;
-  font-size: 26px;
+  font-size: 2.6rem;
   line-height: 1.36;
 `;
 export const Title5 = styled.p`
   font-weight: 700;
-  font-size: 20px;
+  font-size: 2rem;
   line-height: 1.36;
 `;
 export const Title6 = styled.strong`
   font-weight: 700;
-  font-size: 16px;
+  font-size: 1.6rem;
   line-height: 1.36;
 `;
 
 export const Body1 = styled.p`
   font-weight: 400;
-  font-size: 16px;
+  font-size: 1.6rem;
   line-height: 1.44;
 `;
 
 export const Body2 = styled.p`
   font-weight: 700;
-  font-size: 14px;
+  font-size: 1.4rem;
   line-height: 1.44;
 `;
 
 export const Body3 = styled.p`
   font-weight: 600;
-  font-size: 14px;
+  font-size: 1.4rem;
   line-height: 1.44;
 `;
 export const Body4 = styled.p`
   font-weight: 400;
-  font-size: 14px;
+  font-size: 1.4rem;
   line-height: 1.44;
 `;
 
 export const Caption1 = styled.p`
   font-weight: 400;
-  font-size: 12px;
+  font-size: 1.2rem;
   line-height: 1.28;
 `;
 
 export const Caption2 = styled.p`
   font-weight: 400;
-  font-size: 10px;
+  font-size: 1rem;
   line-height: 1.28;
 `;
